fix: move components to /components directory
