+++ src/pages/integration/index.jsx
@@ -1,2 +1,2 @@
 export { default as Integrations } from './IntegrationIndex/IntegrationIndexPage';
-export { default as WidgetPreview } from './WidgetPreview';
+export { default as WidgetPreview } from '../../components/integration/WidgetPreview';
